backend/requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.2
attrs==23.2.0
aws-cdk-lib==2.122.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.122.0a0
aws-psycopg2==1.3.8
aws-wsgi==0.2.7
awsebcli==3.20.10
blinker==1.6.2
boto3==1.34.33
botocore==1.31.0
cachetools==5.3.1
cattrs==23.2.3
cement==2.8.2
certifi==2022.9.24
cffi==1.16.0
cfn-flip==1.3.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
constructs==10.3.0
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.8
distro==1.9.0
durationpy==0.6
et-xmlfile==1.1.0
filelock==3.13.1
Flask==2.3.2
flask-lambda==0.0.4
flask-lambda2==2.0.0
fonttools==4.38.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.62.0
gspread==5.12.1
gspread-formatting==1.1.2
h11==0.14.0
hjson==3.1.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsii==1.94.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kappa==0.6.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
mdurl==0.1.2
MouseInfo==0.1.3
numpy==1.24.1
oauthlib==3.2.2
openai==1.14.2
openpyxl==3.1.2
optional-django==0.3.0
outcome==1.3.0.post0
packaging==22.0
pandas==1.5.2
pathspec==0.10.1
Pillow==9.4.0
placebo==0.9.0
platformdirs==4.1.0
plotly==5.11.0
protobuf==4.25.2
psycopg2==2.9.5
psycopg2-binary==2.9.9
publication==0.0.3
pyarrow==16.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.3
pytweening==1.2.0
pytz==2022.7
pywin32==306
PyYAML==6.0.1
react==4.3.0
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.0
selenium==4.19.0
semantic-version==2.8.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
streamlit==1.35.0
tenacity==8.1.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
trio==0.25.0
trio-websocket==0.11.1
troposphere==4.5.3
typeguard==2.13.3
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.25.0
watchdog==4.0.1
wcwidth==0.1.9
Werkzeug==2.3.4
wsproto==1.2.0
xmltodict==0.13.0
zappa==0.58.0
